feat(checkout): refactor checkout page to utilize useCheckout hook for payment processing
